+++ src/actors/mod.rs
@@ -1,6 +1,4 @@
 //! Helper actors
 
 pub mod signal;
-
-#[cfg(unix)]
-pub mod dns;
+pub mod resolver;